requirements.txt
pandas>=2.0
gspread>=5.12
google-auth>=2.23
gspread-dataframe>=3.3
openpyxl>=3.1
requests>=2.28
python-dotenv>=1.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
selenium>=4.15
webdriver-manager>=4.0
pillow>=10.0
